CHROMIUM: PM / devfreq: rk3399_dmc: Add new dmcfreq governor
This adds an interrupt based governor for the dmc devfreq device. This
allows faster response for DDR transitions and removes the polling
loop.

It also directly scales the frequency based on usage when it
increases. This is because target DDR load is much smaller than the
peak memory usage.

There are three settings for the new governor. Target load,
hysteresis, and down throttle ms. The target load is used when scaling
the frequency. Hysteresis modifies the interrupt thresholds. The floor
threshold is the target load minus hysteresis times the next lower
frequency. The top threshold is the target load plus hysteresis times
the current frequency. Down throttle ms is the delay between
increasing the frequency and when the frequency can be lowered again.

BUG=b:200250574
TEST=boot kevin and run:
cat /sys/class/devfreq/memory-controller/trans_stat
check that there are ddr transitions happening based on the output
TEST=octane, other perf checks

Signed-off-by: Derek Basehore <dbasehore@chromium.org>

[chromeos-5.10(briannorris):
Conflicts:
	drivers/devfreq/Kconfig
	drivers/devfreq/event/rockchip-dfi.c
	drivers/devfreq/rk3399_dmc.c
	include/soc/rockchip/rk3399_dmc.h
- Drop rk3399_dmc.h (there isn't much shared between drivers)
- Switch to using
   dev_set_drvdata(&data->edev->dev, data->devfreq)
  to expose less from 1 driver to the other
- Several trivial context conflicts
- Account for moved profile struct: "PM / devfreq: rk3399_dmc: Avoid
  static (reused) profile"
- Pull in target load, etc., settings that were in
  https://chromium-review.googlesource.com/431941
- Drop sysfs files; these weren't used anywhere and aren't fully
  bounds-checked. Less code is nice.
- Add new DEVFREQ_GOV_FLAG_IRQ_DRIVEN, to ensure framework doesn't try
  to queue meaningless work items
- SPDX license header

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: I2706d7d25ba64d2eb42bab9577f54d70888b630a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3422176
Reviewed-by: Derek Basehore <dbasehore@chromium.org>