CHROMIUM: PCI: Add double-reset quirk for Marvell 8997
Marvell 8997 PCIe modules may not reinitialize correctly if we only
reset (FLR) once. To work properly, we need to perform one FLR,
reinitialize the device, then perform another FLR. We only have a rough
explanation from the vendor as to why this is necessary, but the essence
seems to be that firmware init after a single FLR doesn't configure
something quite the same as if we perform a second FLR.

This approach works by implementing a callback for
pci_dev_specific_reset() that matches the 8997 device ID under both
Marvell vendor IDs. This reset method performs two resets, restoring the
driver in between.

Technically, this could probably be fixed with an updated firmware, but
the vendor has been MIA for >8 months, so we're not hopeful.

***

5.10 notes: cherry-pick from kernel 4.4 was clean, aside from
(a) trivial context conflicts
(b) just assume the answer is "yes" when probing for reset capabilities,
    instead of bothering to expose pcie_has_flr(); we know the hardware
    by ID already
(c) drop unused 'u32 cap'

Still don't bother submitting upstream, because it's an ugly workaround
for bad firmware.

BUG=b:64074778
TEST=`tast run ... wifi.Reset` on Kevin
TEST=C=0 ; while : ; do
    echo "C=$C $(date)"
    let C+=1
    echo 1 > /sys/kernel/debug/mwifiex/mlan0/reset
    rmmod mwifiex_pcie
    sleep 6
    modprobe mwifiex_pcie
    sleep 3
  done

Change-Id: I0adeb973a69a6f9daa0f22d63dd076ceb15af66f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3402154
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>