BACKPORT: cfg80211: Add support to configure SAE PWE value to drivers
Add support to configure SAE PWE preference from userspace to drivers in
both AP and STA modes. This is needed for cases where the driver takes
care of Authentication frame processing (SME in the driver) so that
correct enforcement of the acceptable PWE derivation mechanism can be
performed.

The userspace applications can pass the sae_pwe value using the
NL80211_ATTR_SAE_PWE attribute in the NL80211_CMD_CONNECT and
NL80211_CMD_START_AP commands to the driver. This allows selection
between the hunting-and-pecking loop and hash-to-element options for PWE
derivation. For backwards compatibility, this new attribute is optional
and if not included, the driver is notified of the value being
unspecified.

Signed-off-by: Rohan Dutta <drohan@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Link: https://lore.kernel.org/r/20201027100910.22283-1-jouni@codeaurora.org
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 9f0ffa418483938d25a15f6ad3891389f333bc59)

Conflicts:
   include/uapi/linux/nl80211.h
   	(contextual conflict)
   net/wireless/nl80211.c
   	(contextual conflict)

BUG=b:247539831
TEST=verify via the steps below
* sudo -u wpa -g wpa wpa_cli set_network 0 mode 2
* sudo -u wpa -g wpa wpa_cli set_network 0 frequency 5180
* sudo -u wpa -g wpa wpa_cli set_network 0 ssid "TAST_TEST_YHwSOcVRCIr2WZaDLdiS"
* sudo -u wpa -g wpa wpa_cli set_network 0 key_mgmt SAE
* sudo -u wpa -g wpa wpa_cli set_network 0 psk "chromeos"
* sudo -u wpa -g wpa wpa_cli set_network 0 proto RSN
* sudo -u wpa -g wpa wpa_cli set_network 0 pairwise CCMP
* sudo -u wpa -g wpa wpa_cli set_network 0 group CCMP
* sudo -u wpa -g wpa wpa_cli set_network 0 ieee80211w 2
* sudo -u wpa -g wpa wpa_cli select_network 0
* run in loop `sudo -u wpa -g wpa wpa_cli status` and wait for `wpa_state=COMPLETE`

Change-Id: Ic6a0eb6d8d76418c3ca26aa050eb639d047c12c7
Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4129283
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Jacek Siuda <jsiuda@google.com>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>