UPSTREAM: bus: mhi: core: Prevent sending multiple RDDM entry callbacks
If an mhi_power_down() is initiated after the device has entered
RDDM and a status callback was provided for it, it is possible
that another BHI interrupt fires while waiting for the MHI
RESET to be cleared. If that happens, MHI host would have moved
a "disabled" execution environment and the check to allow sending
an RDDM status callback will pass when it is should not. Add a
check to see if MHI is in an active state before proceeding.

Signed-off-by: Bhaumik Bhatt <bbhatt@codeaurora.org>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
(cherry picked from commit 0c76b3fa580da8f65333b8774205f0a72ffe844c)

BUG=b:214740367
TEST=Verified wifi

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: If8a2e7309495dd914680ee92142e6f699c65f756
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3393474
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>