UPSTREAM: sched: highmem: Store local kmaps in task struct
Instead of storing the map per CPU provide and use per task storage. That
prepares for local kmaps which are preemptible.

The context switch code is preparatory and not yet in use because
kmap_atomic() runs with preemption disabled. Will be made usable in the
next step.

The context switch logic is safe even when an interrupt happens after
clearing or before restoring the kmaps. The kmap index in task struct is
not modified so any nesting kmap in an interrupt will use unused indices
and on return the counter is the same as before.

Also add an assert into the return to user space code. Going back to user
space with an active kmap local is a nono.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lore.kernel.org/r/20201118204007.372935758@linutronix.de
(cherry picked from commit 5fbda3ecd14a5343644979c98d6eb65b7e7de9d8)

BUG=b:222221770
BUG=b:195001087
BUG=b:191366271
TEST=for arch in amd64-generic arm64-generic; do \
        USE=buildtest emerge-"${arch}" chromeos-kernel-5_10; \
     done

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: I467b23aa5be91ec4adda3cb3ae7bcce59661715f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3500321
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>