UPSTREAM: PCI/ERR: Use "bridge" for clarity in pcie_do_recovery()
pcie_do_recovery() may be called with "dev" being either a bridge (Root
Port or Switch Downstream Port) or an Endpoint.  The bulk of the function
deals with the bridge, so if we start with an Endpoint, we reset "dev" to
be the bridge leading to it.

For clarity, replace "dev" in the body of the function with "bridge".  No
functional change intended.

Link: https://lore.kernel.org/r/20201121001036.8560-8-sean.v.kelley@intel.com
Tested-by: Jonathan Cameron <Jonathan.Cameron@huawei.com> # non-native/no RCEC
Signed-off-by: Sean V Kelley <sean.v.kelley@intel.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>
Acked-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
(cherry-picked from commit 0791721d800790e6e533bd8467df67f0dc4f2fec)

BUG=b:217419041
TEST=Connect Gatkex card and authorize it & run suspend_stress_test -c 10  . S0ix test passes 10 cycles

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Ieaa0b6604588e8506bea9b21f144f0e5644259d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3459472
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rajat Jain <rajatja@google.com>