CHROMIUM: iwl7000: pcie: work around ROM bug on AX210 integrated
On 22000 and AX210 devices, there's a ROM bug that causes it to
set invalid LTR settings. On 22000 and AX210 non-integrated we
can fix up these settings from the driver (as done in the code
here), but on AX210 integrated these registers are not available
to the driver.

Attempt to work around the issue by spinning while the IML is
being loaded, the IML will then reprogram the LTR values itself
after it's loaded, so only the brief IML load (which the ROM is
doing) is affected.

(which doesn't make much sense since that's a discrete AX210??)

Change-Id: If08da23e960b6236f8c05c06fc8b26041ac89f1e
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: fef329cfff203c1f99314c2100a2c287578c5106