UPSTREAM: scsi: ufs: Refine error history functions
The UFS error history does not only have "history of errors" but also a
log of some other events which are not defined as errors.

This patch fixes the confused naming of related functions and changes the
approach for updating and printing history in preparation of next patch.

This patch does not change any functionality.

Link: https://lore.kernel.org/r/20201205115901.26815-3-stanley.chu@mediatek.com
Reviewed-by: Asutosh Das <asutoshd@codeaurora.org>
Reviewed-by: Can Guo <cang@codeaurora.org>
Signed-off-by: Stanley Chu <stanley.chu@mediatek.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit e965e5e00b23c47b7a9834436972aa37c5baa708)

BUG=b:245912634
TEST=build and check driver bind

Signed-off-by: Mohandass, Divagar <divagar.mohandass@intel.com>
Change-Id: I63a21e78ec8c370cb16a04d787a849d51cbef444
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3889129
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>