UPSTREAM: crypto: aesni - implement support for cts(cbc(aes))
Follow the same approach as the arm64 driver for implementing a version
of AES-NI in CBC mode that supports ciphertext stealing. This results in
a ~2x speed increase for relatively short inputs (less than 256 bytes),
which is relevant given that AES-CBC with ciphertext stealing is used
for filename encryption in the fscrypt layer. For larger inputs, the
speedup is still significant (~25% on decryption, ~6% on encryption)

Tested-by: Eric Biggers <ebiggers@google.com> # x86_64
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit ddf169a98f01d6fd46295ec0dd4c1d6385be65d4)

BUG=b:198473496
TEST=compilation, tcrypt tests

Signed-off-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Change-Id: I9035b29290bac8e0e5bccd73a790ef1c71db3ea6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3399409
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>