UPSTREAM: media: ext-ctrls-codec-stateless.rst: fix an H-264 table format
The Picture Parameter Set Flags table for H-264 doesn't output
well on PDF. There are missing format instructions for LaTeX, and
the columns are too long.

Reduce the size of the second column, as 16 bits is more than enough
for the current flags usage, and add the needed bits for it to be
properly output in PDF format.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 86d3a055a46749eaf3c15c3cba8b8a137de7073a)

BUG=b:250824543
TEST=Build for Cherry

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I5a314d1bf8c83392feeec1586d1bdceeb00e4aab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3951732
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Steve Cho <stevecho@chromium.org>