CHROMIUM: Bluetooth: Ignore global Connectable setting for extended adv
Existing logic enforces a "connectable" advertisement if the global
Connectable setting is active, even if the advertisement requested was
non-connectable. This can cause a case where the client requires a
random address, but kernel will ignore that request. This patch removes
the check for the global setting, to ensure we provide a random address
if the client did not specify Connectability in their advertisement
request.

Specifically, this seems to break bluetooth discovery on hatch, as
Nearby immediately requests an advertisement on login. The global
Connectable setting makes this advertisement public (not good), and then
future discovery requests fail because the random address update is
skipped since adv is enabled.

Change was tested by verifying Nearby startup flow, and ensuring
both the scanning and advertising steps succeed on hatch.

BUG=b:175069101, b:197529377
TEST=Test Nearby startup flow on hatch, b/197529377#comment22

Signed-off-by: Daniel Winkler <danielwinkler@google.com>
Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Change-Id: I99ea32b02b50ec996239e1a4e489cba47ae90a62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3507074
Reviewed-by: Alain Michaud <alainm@chromium.org>