UPSTREAM: ASoC: wm_adsp: Improve handling of raw byte streams
As the register map is 16-bit or 32-bit big-endian, the 24-bit
DSP words appear padded and with the bytes swapped. When reading a
raw stream of bytes, the pad bytes must be removed and the data bytes
swapped back to their original order.

The previous implementation of this assumed that the be32_to_cpu() in
wm_adsp_read_data_block() would swap back to little-endian. But this is
obviously only true on a little-endian CPU. It also made two walks
through the data, once to endian-swap and again to strip the pad bytes.

This patch re-works the code so that the endian-swap and unpad are done
together in a single walk, and it is not dependent on the endianness of
the CPU. The data_word_size argument to wm_adsp_remove_padding() has been
dropped because currently this is always 3.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20201216112512.26503-1-rf@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 7726e49837af634accaec317c8d246d1d90d8fc5)

BUG=b:186972464
TEST=ODM tests audio function on Vell

Signed-off-by: Brent Lu <brent.lu@intel.com>
Change-Id: Ife692f349afa94d7fc1506ccef3c1bd8ffbdd4e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3471200
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Shon Wang <shon.wang@quanta.corp-partner.google.com>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>