UPSTREAM: mhi: pci_generic: Add health-check
If the modem crashes for any reason, we may not be able to detect
it at MHI level (MHI registers not reachable anymore).

This patch implements a health-check mechanism to check regularly
that device is alive (MHI layer can communicate with). If device
is not alive (because a crash or unexpected reset), the recovery
procedure is triggered.

Tested successfully with Telit FN980m module.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Reviewed-by: Hemant Kumar <hemantk@codeaurora.org>
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
(cherry picked from commit 8562d4fe34a3ef52da077f77985994bb9ad1f83e)

BUG=b:214740367
TEST=Verified wifi

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I75b76af95bb828c907b53b98856ff9d2989ec4c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3393492
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>