BACKPORT: UPSTREAM: swiotlb: Validate bounce size in the sync/unmap path
The size of the buffer being bounced is not checked if it happens
to be larger than the size of the mapped buffer. Because the size
can be controlled by a device, as it's the case with virtio devices,
this can lead to memory corruption.

This patch saves the remaining buffer memory for each slab and uses
that information for validation in the sync/unmap paths before
swiotlb_bounce is called.

Validating this argument is important under the threat models of
AMD SEV-SNP and Intel TDX, where the HV is considered untrusted.

Signed-off-by: Martin Radev <martin.b.radev@gmail.com>
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>

(cherry picked from commit daf9514fd5eb098d7d6f3a1247cb8cc48fc94155)
[downstream: solve conflict in swiotlb_tbl_sync_single()]

BUG=b:202792011
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: Ia73140d77b45336769f7afaf83936a55f89be042
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3251259
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>