UPSTREAM: media: ov9640: Use the generic clock framework
Commit 63839882c597 ("media: mach-pxa: palmz72/pcm990: remove soc_camera dependencies")
removed the last in-tree user of this sensor. New users
will be required to use the generic clock framework,
so it's possible to convert the driver to use it.

Convert the driver to use the CCF, and drop the legacy
v4l2-clk API.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Acked-by: Petr Cvek <petrcvekcz@gmail.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 8de14b3acb9c7f4bcb8565a036dd579e5b7aaa7a)

BUG=b:250824543
TEST=Build for Cherry and arm64-generic

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I1de48b0c24061c51ccb74d8cad153b6b319d8e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3954877
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>