UPSTREAM: media: ov6650: Use the generic clock framework
Commit ce548396a433 ("media: mach-omap1: board-ams-delta.c: remove soc_camera dependencies")
removed the last in-tree user of this sensor. New users
will be required to use the generic clock framework,
so it's possible to convert the driver to use it.

Convert the driver to use the CCF, and drop the legacy
v4l2-clk API.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Reviewed-by: Janusz Krzysztofik <jmkrzyszt@gmail.com>
Reviewed-by: Petr Cvek <petrcvekcz@gmail.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 01747ab9ca07c086f253bb99ba261def9dfb41d0)

BUG=b:250824543
TEST=Build for Cherry and arm64-generic

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I4d3729a15edf6c9ce714211e089c56bb087d7d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3954879
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>