UPSTREAM: platform/x86: intel-vbtn: Create 2 separate input-devs for buttons and switches
Create 2 separate input-devs for buttons and switches, this is a
preparation for dynamically registering the switches-input device
for devices which are not on the switches allow-list, but do make
Notify() calls with an event value from the switches sparse-keymap.

This also brings the intel-vbtn driver inline with the intel-hid
driver which is doing the same thing.

Cc: Elia Devito <eliadevito@gmail.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://lore.kernel.org/r/20210115161850.117614-2-hdegoede@redhat.com
(cherry picked from commit 034b8c2e7b06777775c55cd2db2b6a98f4791b5f)

BUG=b:233114312
TEST=Basic boot on skolas

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Ie44d014da5d4b34e9f32caa61cc3dc2b1ccc44da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3699042
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Lalith Rajendran <lalithkraj@google.com>
Tested-by: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>