UPSTREAM: platform/x86: intel-vbtn: Add alternative method to enable switches
Some 2-in-1s have a broken VGBS method, so we cannot get an initial
state for the switches from them. Reporting the wrong initial state for
SW_TABLET_MODE causes serious problems (touchpad and/or keyboard events
being ignored by userspace when reporting SW_TABLET_MODE=1), so on these
devices we cannot register an input-dev for the switches at probe time.

We can however register an input-dev for the switches as soon as we
receive the first switches event, because then we will know the state.

Note this mirrors the behavior of recent changs to the intel-hid driver
which also registers a separate switches input-dev on receiving the
first event on machines with a broken VGBS method.

Cc: Elia Devito <eliadevito@gmail.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://lore.kernel.org/r/20210115161850.117614-3-hdegoede@redhat.com
(cherry picked from commit 3a2f53cd03101f6a7cc34c558b0dbfbaca798165)

BUG=b:233114312
TEST=Basic boot on skolas

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Ica54bacc3f0d6e970e7c17741dc9b3385fdcc4d4
Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3699043
Tested-by: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Lalith Rajendran <lalithkraj@google.com>