UPSTREAM: crypto: aesni - release FPU during skcipher walk API calls
Taking ownership of the FPU in kernel mode disables preemption, and
this may result in excessive scheduling blackouts if the size of the
data being processed on the FPU is unbounded.

Given that taking and releasing the FPU is cheap these days on x86, we
can limit the impact of this issue easily for skcipher implementations,
by moving the FPU begin/end calls inside the skcipher walk processing
loop. Considering that skcipher walks operate on at most one page at a
time, doing so fully mitigates this issue.

This also permits the skcipher walk logic to use non-atomic kmalloc()
calls etc so we can change the 'atomic' bool argument in the calls to
skcipher_walk_virt() to false as well.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 65d1e3c415f6e380f6168faf333a59ec235eac5d)

BUG=b:198473496
TEST=compilation, tcrypt tests

Signed-off-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Change-Id: I667918bf7200db57b309263ccddcd00633f6883d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3399416
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>