UPSTREAM: apparmor: handle idmapped mounts
The i_uid and i_gid are mostly used when logging for AppArmor. This is
broken in a bunch of places where the global root id is reported instead
of the i_uid or i_gid of the file. Nonetheless, be kind and log the
mapped inode if we're coming from an idmapped mount. If the initial user
namespace is passed nothing changes so non-idmapped mounts will see
identical behavior as before.

Link: https://lore.kernel.org/r/20210121131959.646623-26-christian.brauner@ubuntu.com
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>
(cherry picked from commit 3cee6079f62f4d3a37d9dda2e0851677e08028ff)

BUG=b:271154170
TEST=CQ

Change-Id: Ia659548a2d7af866e2dfe1aacfd3607dd97f3e52
Signed-off-by: Ben Scarlato <akhna@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4384204
Reviewed-by: Jeff Xu <jeffxu@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Allen Webb <allenwebb@google.com>