UPSTREAM: platform/surface: aggregator: Fix braces in if condition with unlikely() macro
The braces of the unlikely() macro inside the if condition only cover
the subtraction part, not the whole statement. This causes the result of
the subtraction to be converted to zero or one. While that still works
in this context, it causes static analysis tools to complain (and is
just plain wrong).

Fix the bracket placement and, while at it, simplify the if-condition.
Also add a comment to the if-condition explaining what we expect the
result to be and what happens on the failure path, as it seems to have
caused a bit of confusion.

This commit should not cause any difference in behavior or generated
code.

Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Fixes: c167b9c7e3d6 ("platform/surface: Add Surface Aggregator subsystem")
Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>
Link: https://lore.kernel.org/r/20210126172202.1428367-1-luzmaximilian@gmail.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit 2691d0ae668ab9d9f3f275ac6ed6029862780084)

BUG=b:233114312
TEST=Basic boot on skolas

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Icb789c7e0d6970a0369b1deca7c143cd1aa37399
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3733731
Owners-Override: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Lalith Rajendran <lalithkraj@google.com>
Tested-by: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Lalith Rajendran <lalithkraj@google.com>