UPSTREAM: xen-swiotlb: remove xen_set_nslabs
The xen_set_nslabs function is a little weird, as it has just one
caller, that caller passes a global variable as the argument,
which is then overriden in the function and a derivative of it
returned.  Just add a cpp symbol for the default size using a readable
constant and open code the remaining three lines in the caller.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>

(cherry picked from commit 4035b43da6daa51668830becfe8537841414946b)

BUG=b:202792011
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I83d5889f340dad4e9411cb557a7e1ee1ee750a2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3252973
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>