BACKPORT: iommu/dma: Resurrect the "forcedac" option
In converting intel-iommu over to the common IOMMU DMA ops, it quietly
lost the functionality of its "forcedac" option. Since this is a handy
thing both for testing and for performance optimisation on certain
platforms, reimplement it under the common IOMMU parameter namespace.

For the sake of fixing the inadvertent breakage of the Intel-specific
parameter, remove the dmar_forcedac remnants and hook it up as an alias
while documenting the transition to the new common parameter.

Fixes: c588072bba6b ("iommu/vt-d: Convert intel iommu driver to the iommu ops")
Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Acked-by: Lu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: John Garry <john.garry@huawei.com>
Link: https://lore.kernel.org/r/7eece8e0ea7bfbe2cd0e30789e0d46df573af9b0.1614961776.git.robin.murphy@arm.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>

(cherry picked from commit 3542dcb15cef66c0b9e6c3b33168eb657e0d9520)

Conflicts:
   drivers/iommu/intel/iommu.c
[ Preserve the FROMLIST patch.  See b/228395365#comment2. ]

BUG=b:228395365
BUG=b:195152713
BUG=b:184662291
TEST=for arch in amd64-generic arm64-generic; do \
	USE=buildtest emerge-"${arch}" chromeos-kernel-5_10; \
     done

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: I6d3b93f98f2686ceb5f4f91b0297914ac1b2f683
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3583296
Reviewed-by: Guenter Roeck <groeck@chromium.org>