UPSTREAM: platform/x86: intel_pmc_core: Update Kconfig
The intel_pmc_core driver is mostly used as a debugging driver for Intel
platforms that support SLPS0 (S0ix). But the driver may also be used to
communicate actions to the PMC in order to ensure transition to SLPS0 on
some systems and architectures. As such the driver should be built on all
platforms it supports. Indicate this in the Kconfig. Also update the list
of supported features.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Suggested-by: Mario Limonciello <mario.limonciello@dell.com>
Link: https://lore.kernel.org/r/20210319201844.3305399-1-david.e.box@linux.intel.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit a01be40c6313d3bb6e121820847c6650c358be88)

BUG=b:233114312
TEST=Basic boot on skolas

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: Ie15db8539e408ab528834f9bddf8af89bdc023a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3699432
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Lalith Rajendran <lalithkraj@google.com>