UPSTREAM: intel: remove checker warning
The sparse checker (C=2) found an assignment where we were mixing
types when trying to convert from data read directly from the
device NVM, to an array in CPU order in-memory, which
unfortunately the driver tries to do in-place.

This is easily solved by using the swap operation instead of an
assignment, and is already proven in other Intel drivers to be
functionally correct and the same code, just without a sparse
warning.

The change is the same in all three drivers.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Tested-by: Dave Switzer <david.switzer@intel.com>
Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>

(cherry picked from commit c40591cc3d48194faa80bda652d86a1ed8e221be)

BUG=b:195336604
TEST=Connectivity/functional testing on ADL-M RVP. i225 device with FW 1.82. Full tests info: https://docs.google.com/spreadsheets/d/1iFsjZKgyOn93eRs8uTgLnHBCZf0e5IDRM2ezm-0N3q4/edit#gid=127102956

Signed-off-by: Amir Hanania <amir.hanania@intel.corp-partner.google.com>
Change-Id: Ida5c2c336ccde459d41f584ddf0d04ef09841fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3497853
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>