UPSTREAM: bus: mhi: core: Rely on accurate method to determine EDL mode
Relying on the current execution environment to determine if EDL
image was downloaded should not be done as the execution
environment can change at this point in error cases and we may
misread it. Instead, MHI can rely on the local 'fw_name' variable
as a safer alternative.

Signed-off-by: Bhaumik Bhatt <bbhatt@codeaurora.org>
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Reviewed-by: Loic Poulain <loic.poulain@linaro.org>
Link: https://lore.kernel.org/r/1617067704-28850-2-git-send-email-bbhatt@codeaurora.org
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
(cherry picked from commit 418bec695696abd1b4d5abca85f143ead2ffedac)

BUG=b:214740367
TEST=Verified wifi

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: Ia600814bf314efc1b1e4b5226ed90cbc5f51e06d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3393520
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>