UPSTREAM: usb: typec: Port mapping utility
Adding functions that can be used to link/unlink ports -
USB ports, TBT3/USB4 ports, DisplayPorts and so on - to
the USB Type-C connectors they are attached to inside a
system. The symlink that is created for the port device is
named "connector".

Initially only ACPI is supported. ACPI port object shares
the _PLD (Physical Location of Device) with the USB Type-C
connector that it's attached to.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Link: https://lore.kernel.org/r/20210407065555.88110-2-heikki.krogerus@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit ae196ddb0d3186bc08e529b8ea4bf62161ddfce2)

BUG=b:202446737
TEST=Verify that connected USB devices point to the right Type C port
     device.

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I18c50793841530f6b1e80468711ee7f008b2bc13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3213459
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>