UPSTREAM: r8152: set inter fram gap time depending on speed
Set the maximum inter frame gap time (144ns) for speed 10M/half and
100M/half. It improves the performance for those speeds. And, there
is no effect for the other speeds.

For 10M/half and 100M/half, the fast inter frame gap time let the
device couldn't use the feature of the aggregation effectively,
because the transfer would be completed fastly. Therefore, use the
maximum value to improve the effect of the aggregation. However, you
may not feel the improvement for fast CPUs, because they compensate
for the effect of the aggregation.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 5133bcc7481528e36fff0a3b056601efb704fb32)

BUG=b:198239768
TEST=emerge-amd64-generic --nodeps chromeos-kernel-5_10

Signed-off-by: Grant Grundler <grundler@chromium.org>
Change-Id: I7cee2066b5eb081449a1d7c8306132244db7c8fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3145099
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>