UPSTREAM: landlock: Add ruleset and domain management
A Landlock ruleset is mainly a red-black tree with Landlock rules as
nodes.  This enables quick update and lookup to match a requested
access, e.g. to a file.  A ruleset is usable through a dedicated file
descriptor (cf. following commit implementing syscalls) which enables a
process to create and populate a ruleset with new rules.

A domain is a ruleset tied to a set of processes.  This group of rules
defines the security policy enforced on these processes and their future
children.  A domain can transition to a new domain which is the
intersection of all its constraints and those of a ruleset provided by
the current process.  This modification only impact the current process.
This means that a process can only gain more constraints (i.e. lose
accesses) over time.

Cc: James Morris <jmorris@namei.org>
Signed-off-by: Mickaël Salaün <mic@linux.microsoft.com>
Acked-by: Serge Hallyn <serge@hallyn.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Jann Horn <jannh@google.com>
Link: https://lore.kernel.org/r/20210422154123.13086-3-mic@digikod.net
Signed-off-by: James Morris <jamorris@linux.microsoft.com>
(cherry picked from commit ae271c1b14de343b888e77f74f640e3dcbdeb4c9)

BUG=b:232966959
TEST=test landlock sample

Signed-off-by: Jeff Xu <jeffxu@google.com>
Change-Id: I17b270e198252c53582bb409a5c8ac8ce593f93d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3710334
Reviewed-by: Mickaël Salaün <mic@digikod.net>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>