BACKPORT: landlock: Set up the security framework and manage credentials
Process's credentials point to a Landlock domain, which is underneath
implemented with a ruleset.  In the following commits, this domain is
used to check and enforce the ptrace and filesystem security policies.
A domain is inherited from a parent to its child the same way a thread
inherits a seccomp policy.

Cc: James Morris <jmorris@namei.org>
Signed-off-by: Mickaël Salaün <mic@linux.microsoft.com>
Reviewed-by: Jann Horn <jannh@google.com>
Acked-by: Serge Hallyn <serge@hallyn.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Link: https://lore.kernel.org/r/20210422154123.13086-4-mic@digikod.net
Signed-off-by: James Morris <jamorris@linux.microsoft.com>

(cherry picked from commit 385975dca53eb41031d0cbd1de318eb1bc5d6bb9)

Conflicts:
   security/Kconfig

BUG=b:232966959
TEST=test landlock sample

Signed-off-by: Jeff Xu <jeffxu@google.com>
Change-Id: Idaf9a36de171fb8bce94b8dac2ed097c1c7f50cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3710335
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mickaël Salaün <mic@digikod.net>
Reviewed-by: Guenter Roeck <groeck@chromium.org>