UPSTREAM: landlock: Add ptrace restrictions
Using ptrace(2) and related debug features on a target process can lead
to a privilege escalation.  Indeed, ptrace(2) can be used by an attacker
to impersonate another task and to remain undetected while performing
malicious activities.  Thanks to  ptrace_may_access(), various part of
the kernel can check if a tracer is more privileged than a tracee.

A landlocked process has fewer privileges than a non-landlocked process
and must then be subject to additional restrictions when manipulating
processes. To be allowed to use ptrace(2) and related syscalls on a
target process, a landlocked process must have a subset of the target
process's rules (i.e. the tracee must be in a sub-domain of the tracer).

Cc: James Morris <jmorris@namei.org>
Signed-off-by: Mickaël Salaün <mic@linux.microsoft.com>
Reviewed-by: Jann Horn <jannh@google.com>
Acked-by: Serge Hallyn <serge@hallyn.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Link: https://lore.kernel.org/r/20210422154123.13086-5-mic@digikod.net
Signed-off-by: James Morris <jamorris@linux.microsoft.com>
(cherry picked from commit afe81f754117dd96853677c5cb815f49abef0ba0)

BUG=b:232966959
TEST=test landlock sample

Signed-off-by: Jeff Xu <jeffxu@google.com>
Change-Id: Idaa586f97a02af27df2b4fe6f0dc04ee756038c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3710336
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Mickaël Salaün <mic@digikod.net>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>