UPSTREAM: media: hantro/cedrus: Remove unneeded slice size and slice offset
The MPEG2_SLICE_PARAMS control is designed to refer to a
single slice. However, the Hantro and Cedrus drivers operate
in per-frame mode, and so does the current Ffmpeg and GStreamer
implementations that are tested with these two drivers.

In other words, the drivers are expecting all the slices in a picture
(with either frame or field structure) to be contained in
the OUTPUT buffer, which means the slice size and offset shouldn't be used.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Tested-by: Jernej Skrabec <jernej.skrabec@siol.net>
Reviewed-by: Jernej Skrabec <jernej.skrabec@siol.net>
Tested-by: Daniel Almeida <daniel.almeida@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 701a6a410c319729c86bfb696860f21adbff1bfa)

BUG=b:250824543
TEST=Build for Cherry

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Ifc221f8e30a07483f7a0a59b41dce4105ccf1fb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3951744
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Steve Cho <stevecho@chromium.org>