UPSTREAM: platform/x86: Add intel_skl_int3472 driver
ACPI devices with _HID INT3472 are currently matched to the tps68470
driver, however this does not cover all situations in which that _HID
occurs. We've encountered three possibilities:

1. On Chrome OS devices, an ACPI device with _HID INT3472 (representing
a physical TPS68470 device) that requires a GPIO and OpRegion driver
2. On devices designed for Windows, an ACPI device with _HID INT3472
(again representing a physical TPS68470 device) which requires GPIO,
Clock and Regulator drivers.
3. On other devices designed for Windows, an ACPI device with _HID
INT3472 which does **not** represent a physical TPS68470, and is instead
used as a dummy device to group some system GPIO lines which are meant
to be consumed by the sensor that is dependent on this entry.

This commit adds a new module, registering a platform driver to deal
with the 3rd scenario plus an i2c driver to deal with #1 and #2, by
querying the CLDB buffer found against INT3472 entries to determine
which is most appropriate.

Suggested-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Daniel Scally <djrscally@gmail.com>
Link: https://lore.kernel.org/r/20210603224007.120560-6-djrscally@gmail.com
[hdegoede@redhat.com Make skl_int3472_tps68470_calc_type() static]
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit 5de691bffe57fd0fc2b4dcdcf13815c56d11db10)

BUG=b:233114312
TEST=Basic boot on skolas

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: I6636ce974aee88ae5221af483b15b5784fa2c5f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3711251
Reviewed-by: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Lalith Rajendran <lalithkraj@google.com>
Commit-Queue: Lalith Rajendran <lalithkraj@google.com>