UPSTREAM: drm/tegra: Allocate per-engine channel in core code
To avoid code duplication, allocate the per-engine shared channel in
the core code instead. This is the usual channel that all jobs are
submitted to when MLOCKing is not in use. Once MLOCKs are implemented
on Host1x side, we can also update this to avoid allocating a shared
channel when MLOCKs are enabled.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit e0f2977c35733fcccafdabcc02bf4f6ddf7f438d)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:207759642
TEST=Tested on brya/zork/tomato/trogdor/grunt

Backported using forklift.py

Change-Id: I7472b7806d1f66c2b5ec5b8598c814694993e198