UPSTREAM: drm/tegra: Add job firewall
Add a firewall that validates jobs before submission to ensure
they don't do anything they aren't allowed to do, like accessing
memory they should not access.

The firewall is functionality-wise a copy of the firewall already
implemented in gpu/host1x. It is copied here as it makes more
sense for it to live on the DRM side, as it is only needed for
userspace job submissions, and generally the data it needs to
do its job is easier to access here.

In the future, the other implementation will be removed.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 8cc95f3fd35e1de3495d9e2acb051162c739cd25)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:207759642
TEST=Tested on brya/zork/tomato/trogdor/grunt

Backported using forklift.py

Change-Id: Ibfabc24999124abbbd28769484c2ef760d0e3d8a